:sparkles: feat: create the Status component
