+++ src/components/profiles/Status/Status.stories.tsx
@@ -1,0 +1,15 @@
+import Status from './Status';
+
+import type { Meta, StoryObj } from '@storybook/react';
+
+const meta: Meta<typeof Status> = {
+  component: Status,
+};
+
+export default meta;
+
+type Story = StoryObj<typeof Status>;
+
+export const Default: Story = {
+  render: () => <Status />,
+};
